big revision of audio examples that closes different issues
Given the announce that a 0.55-1 update was coming soon, I made a big revision that gathered different issues into a single commit, all at once.

So... this:
closes https://github.com/pure-data/pddp/issues/232
closes https://github.com/pure-data/pddp/issues/235
closes https://github.com/pure-data/pddp/issues/237
closes https://github.com/pure-data/pddp/issues/238

I also renamed H16 and simplified just to use [text] instead of [qlist], there's a previous example that is still named and mainly uses [qlist]. I think this is good because [text] is a more modern solution.